Providing PostgreSql rule to run for tests
